Rename file calendar/event.rs -> calendar/calendar_event.rs
